FIx circular dependencies in $lib/server/api/admins
This one required breaking out some logic into smaller modules imported
by both files, rather than having them import from each other.
