+++ src/lib/server/api/core/admins/cache.ts
@@ -1,0 +1,2 @@
+export const redisString = (instance_id: number, admin_id: number | 'all') =>
+	`i:${instance_id}:admin:${admin_id}`;
